Add device remote controlling for tenants
The remote control feature didn't work for tenants and this commit fixes that issue.

The root cause of this issue is that the server doesn't connect to the WebSocket in the tenant mode.

In order to connect to the WebSocket, it is required to provide valid access token. When getting access token, in the tenant mode it is required to pass tenant domain along with the context of the username.